+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Changes
 *******
 
+Dev
+----
+
+- Add staticfiles finder to serve compiled files in dev mode
+
+
 0.6
 ----